Fixes/additions to the unit test running script.

*    Parallelize the test discovery phase, to reduce lag before actually launching the test
*    Allow users to specify multiple unit tests to run on the command, a la integration.py
*    Convert the -1 option from a test specification option to a running mode option (the current ways of using run.py should transfer.)
*    Adjust the mute defaults such that we mute by default unless running a single test.
*    Adjust timeout such that it's turned off when running explicitly specified tests
*    Change things such that you use the same Python for scons path introspection as you're using to run the script.

Old repository SHA1: 968a049903cdd92b0550b4a4a3304957ee5d5882

+++ source/test/timelimit.py
@@ -24,7 +24,7 @@ def main(argv):
         print('Usage: "timelimit <number of minutes to wait> <command line> <args...>"')
         return 1
 
-    timeout = int(argv[1])
+    timeout = float(argv[1])
 
     commline = ' '.join(argv[2:])
     proc = subprocess.Popen(["bash", "-c", commline])
